+++ src/types/transaction-type/index.ts
@@ -1,0 +1,1 @@
+export type TTransactionType = string